X11: fix usage of required configuration values
Before has Kodi checked only the EGL_NATIVE_VISUAL_ID and ignored
others. With this was e.g. on a some addons the needed EGL_DEPTH_SIZE
not available.

This separates a part of IsSuitableVisual (...) into a separate function
(SuitableCheck (...)) which is checked by GetEGLConfig and takes the
necessary configuration from the list.