[windowing/gbm] Select best CRTC+planes option
On a platform with multiple CRTCs connected to the selected encoder, instead of
using the default CRTC currently driving the encoder, check the availability of
video and GUI planes on each one, then select the best out of all possible
CRTC+planes options.

In the case of choosing a non-default CRTC, we need to explicitly disable the
formerly used CRTC for the connector when using the DRM Atomic API to modeset
using the new CRTC, otherwise the test commit will fail with EINVAL. This is not
necessary for the DRM Legacy API as it will disable the former CRTC implicitly.